+++ src/Camera.cpp
@@ -12,7 +12,7 @@ Camera::Camera(glm::vec3 position, glm::vec3 front, glm::vec3 up) {
     pitch = 0.0f;
 
 }
-void Camera::processKeyboardInput(Camera_Movement direction, float deltaTime) {
+void Camera::processKeyboardInput(Camera_Movement direction, double deltaTime) {
     const float cameraSpeed = 150.0f * deltaTime;
 
     glm::vec3 right = glm::normalize(glm::cross(m_front, m_up));
